% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{addseven}
\alias{addseven}
\title{Add seven function}
\usage{
addseven(x)
}
\arguments{
\item{x}{A numeric base value}
}
\description{
This function adds 7 to the number given.
}
\examples{
addseven()
}
\keyword{add}
